Revert "chore: refactor"
This reverts commit b906fea32e8aec22799c7b532b5caf566e47446b.
