+++ apps/web/src/components/Composer/LinkPreviews.tsx
@@ -5,7 +5,7 @@ import { XMarkIcon } from '@heroicons/react/24/outline';
 import { KNOWN_ATTRIBUTES } from '@hey/data/constants';
 import getURLs from '@hey/helpers/getURLs';
 import { MetadataAttributeType } from '@lens-protocol/metadata';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { usePublicationAttachmentStore } from 'src/store/non-persisted/publication/usePublicationAttachmentStore';
 import { usePublicationAttributesStore } from 'src/store/non-persisted/publication/usePublicationAttributesStore';
 import { usePublicationStore } from 'src/store/non-persisted/publication/usePublicationStore';
@@ -15,7 +15,6 @@ const LinkPreviews: FC = () => {
   const { attachments } = usePublicationAttachmentStore((state) => state);
   const { addAttribute, getAttribute, removeAttribute } =
     usePublicationAttributesStore();
-  const [showRemove, setShowRemove] = useState(false);
 
   const urls = getURLs(publicationContent);
 
@@ -37,27 +36,22 @@ const LinkPreviews: FC = () => {
 
   return (
     <div className="relative m-5">
-      <Oembed
-        onLoad={(og) => setShowRemove(og?.title ? true : false)}
-        url={urls[0]}
-      />
-      {showRemove && (
-        <div className="absolute top-0 -m-3">
-          <button
-            className="rounded-full bg-gray-900 p-1.5 opacity-75"
-            onClick={() =>
-              addAttribute({
-                key: KNOWN_ATTRIBUTES.HIDE_OEMBED,
-                type: MetadataAttributeType.BOOLEAN,
-                value: 'true'
-              })
-            }
-            type="button"
-          >
-            <XMarkIcon className="size-4 text-white" />
-          </button>
-        </div>
-      )}
+      <Oembed url={urls[0]} />
+      <div className="absolute top-0 m-3">
+        <button
+          className="rounded-full bg-gray-900 p-1.5 opacity-75"
+          onClick={() =>
+            addAttribute({
+              key: KNOWN_ATTRIBUTES.HIDE_OEMBED,
+              type: MetadataAttributeType.BOOLEAN,
+              value: 'true'
+            })
+          }
+          type="button"
+        >
+          <XMarkIcon className="size-4 text-white" />
+        </button>
+      </div>
     </div>
   );
 };
