+++ apps/web/src/components/Composer/NewPublication.tsx
@@ -17,23 +17,28 @@ import QuotedPublication from '@components/Publication/QuotedPublication';
 import { AudioPublicationSchema } from '@components/Shared/Audio';
 import Wrapper from '@components/Shared/Embed/Wrapper';
 import errorToast from '@helpers/errorToast';
+import getNftOpenActionKit from '@helpers/getNftOpenActionKit';
 import { Leafwatch } from '@helpers/leafwatch';
 import uploadToArweave from '@helpers/uploadToArweave';
 import { KNOWN_ATTRIBUTES } from '@hey/data/constants';
 import { Errors } from '@hey/data/errors';
 import { PUBLICATION } from '@hey/data/tracking';
+import { VerifiedOpenActionModules } from '@hey/data/verified-openaction-modules';
 import checkDispatcherPermissions from '@hey/helpers/checkDispatcherPermissions';
 import collectModuleParams from '@hey/helpers/collectModuleParams';
 import getProfile from '@hey/helpers/getProfile';
+import getURLs from '@hey/helpers/getURLs';
 import removeQuoteOn from '@hey/helpers/removeQuoteOn';
 import { ReferenceModuleType } from '@hey/lens';
 import { Button, Card, ErrorMessage } from '@hey/ui';
 import cn from '@hey/ui/cn';
 import { MetadataAttributeType } from '@lens-protocol/metadata';
+import { useQuery } from '@tanstack/react-query';
 import { useUnmountEffect } from 'framer-motion';
 import dynamic from 'next/dynamic';
 import { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
+import { HEY_REFERRAL_PROFILE_ID } from 'src/constants';
 import useCreatePoll from 'src/hooks/useCreatePoll';
 import useCreatePublication from 'src/hooks/useCreatePublication';
 import usePublicationMetadata from 'src/hooks/usePublicationMetadata';
@@ -168,9 +173,6 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
   const [isLoading, setIsLoading] = useState(false);
   const [showEmojiPicker, setShowEmojiPicker] = useState<boolean>(false);
   const [publicationContentError, setPublicationContentError] = useState('');
-  const [publicationNftOpenActionModule, setPublicationNftOpenActionModule] =
-    useState();
-  const [hasOpenAction, setHasOpenAction] = useState(false);
 
   const editor = useEditorContext();
 
@@ -287,6 +289,41 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
     setPublicationContentError('');
   }, [audioPublication]);
 
+  const fetchnftOpenActionEmbed = async (
+    publicationContent: string
+  ): Promise<any | undefined> => {
+    const nftOpenActionKit = getNftOpenActionKit();
+    const publicationContentUrls = getURLs(publicationContent);
+
+    try {
+      const calldata = await nftOpenActionKit.detectAndReturnCalldata({
+        contentURI: publicationContentUrls[0],
+        publishingClientProfileId: HEY_REFERRAL_PROFILE_ID
+      });
+
+      if (calldata) {
+        return {
+          unknownOpenAction: {
+            address: VerifiedOpenActionModules.DecentNFT,
+            data: calldata
+          }
+        };
+      } else {
+        return undefined;
+      }
+    } catch (error) {
+      console.error('Error fetching open action embed:', error);
+      return undefined;
+    }
+  };
+
+  const { data: nftOpenActionEmbed, isLoading: nftOpenActionEmbedLoading } =
+    useQuery({
+      enabled: Boolean(publicationContent),
+      queryFn: () => fetchnftOpenActionEmbed(publicationContent),
+      queryKey: ['fetchnftOpenActionEmbed', publicationContent]
+    });
+
   const getAnimationUrl = () => {
     const fallback =
       'ipfs://bafkreiaoua5s4iyg4gkfjzl6mzgenw4qw7mwgxj7zf7ev7gga72o5d3lf4';
@@ -380,6 +417,10 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
       // Payload for the open action module
       const openActionModules = [];
 
+      if (Boolean(nftOpenActionEmbed)) {
+        openActionModules.push(nftOpenActionEmbed);
+      }
+
       if (Boolean(collectModule.type)) {
         openActionModules.push({
           collectOpenAction: collectModuleParams(collectModule, currentProfile)
@@ -388,10 +429,6 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
 
       if (Boolean(openAction)) {
         openActionModules.push({ unknownOpenAction: openAction });
-      }
-
-      if (!!publicationNftOpenActionModule) {
-        openActionModules.push(publicationNftOpenActionModule);
       }
 
       // Payload for the Momoka post/comment/quote
@@ -404,7 +441,7 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
         contentURI: `ar://${arweaveId}`
       };
 
-      if (useMomoka && !publicationNftOpenActionModule) {
+      if (useMomoka && !nftOpenActionEmbed) {
         if (canUseLensManager) {
           if (isComment) {
             return await createCommentOnMomka(
@@ -554,11 +591,14 @@ const NewPublication: FC<NewPublicationProps> = ({ publication }) => {
       ) : null}
       {showPollEditor ? <PollEditor /> : null}
       {showLiveVideoEditor ? <LivestreamEditor /> : null}
-      <OpenActions
-        setHasOpenAction={setHasOpenAction}
-        setPublicationNftOpenActionModule={setPublicationNftOpenActionModule}
-      />
-      {!hasOpenAction ? <LinkPreviews /> : null}
+      {Boolean(nftOpenActionEmbed) ? (
+        <OpenActions
+          nftOpenActionEmbed={nftOpenActionEmbed}
+          nftOpenActionEmbedLoading={nftOpenActionEmbedLoading}
+        />
+      ) : (
+        <LinkPreviews />
+      )}
       <NewAttachments attachments={attachments} />
       {quotedPublication ? (
         <Wrapper className="m-5" zeroPadding>
