+++ apps/web/src/components/Composer/OpenActions.tsx
@@ -1,28 +1,25 @@
-import DecentOpenAction from '@components/Publication/OpenAction/UnknownModule/Decent';
 import SwapOpenAction from '@components/Publication/OpenAction/UnknownModule/Swap';
-import getNftOpenActionKit from '@helpers/getNftOpenActionKit';
+import Oembed from '@components/Shared/Oembed';
 import { XMarkIcon } from '@heroicons/react/24/outline';
 import { KNOWN_ATTRIBUTES } from '@hey/data/constants';
 import { VerifiedOpenActionModules } from '@hey/data/verified-openaction-modules';
 import getURLs from '@hey/helpers/getURLs';
 import { type UnknownOpenActionModuleSettings } from '@hey/lens';
 import { MetadataAttributeType } from '@lens-protocol/metadata';
-import { useQuery } from '@tanstack/react-query';
 import { type FC, useEffect } from 'react';
-import { HEY_REFERRAL_PROFILE_ID } from 'src/constants';
 import { useOpenActionStore } from 'src/store/non-persisted/publication/useOpenActionStore';
 import { usePublicationAttachmentStore } from 'src/store/non-persisted/publication/usePublicationAttachmentStore';
 import { usePublicationAttributesStore } from 'src/store/non-persisted/publication/usePublicationAttributesStore';
 import { usePublicationStore } from 'src/store/non-persisted/publication/usePublicationStore';
 
-type OpenActionProps = {
-  setHasOpenAction: (hasOpenAction: boolean) => void;
-  setPublicationNftOpenActionModule: (nftOpenActionEmbed: any) => void;
+type OpenActionsProps = {
+  nftOpenActionEmbed?: boolean;
+  nftOpenActionEmbedLoading?: boolean;
 };
 
-const OpenActions: FC<OpenActionProps> = ({
-  setHasOpenAction,
-  setPublicationNftOpenActionModule
+const OpenActions: FC<OpenActionsProps> = ({
+  nftOpenActionEmbed,
+  nftOpenActionEmbedLoading
 }) => {
   const { openAction, reset } = useOpenActionStore();
 
@@ -32,43 +29,6 @@ const OpenActions: FC<OpenActionProps> = ({
     usePublicationAttributesStore();
 
   const urls = getURLs(publicationContent);
-  const hasSwapOpenAction =
-    openAction?.address === VerifiedOpenActionModules.Swap;
-
-  const fetchnftOpenActionEmbed = async (
-    publicationContent: string
-  ): Promise<any | undefined> => {
-    const nftOpenActionKit = getNftOpenActionKit();
-    const publicationContentUrls = getURLs(publicationContent);
-
-    try {
-      const calldata = await nftOpenActionKit.detectAndReturnCalldata({
-        contentURI: publicationContentUrls[0],
-        publishingClientProfileId: HEY_REFERRAL_PROFILE_ID
-      });
-
-      if (calldata) {
-        return {
-          unknownOpenAction: {
-            address: VerifiedOpenActionModules.DecentNFT,
-            data: calldata
-          }
-        };
-      } else {
-        return undefined;
-      }
-    } catch (error) {
-      console.error('Error fetching open action embed:', error);
-      return undefined;
-    }
-  };
-
-  const { data: nftOpenActionEmbed, isLoading: nftOpenActionEmbedLoading } =
-    useQuery({
-      enabled: Boolean(publicationContent),
-      queryFn: () => fetchnftOpenActionEmbed(publicationContent),
-      queryKey: ['fetchnftOpenActionEmbed', publicationContent]
-    });
 
   useEffect(() => {
     if (urls.length) {
@@ -77,24 +37,10 @@ const OpenActions: FC<OpenActionProps> = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [urls.length]);
 
-  useEffect(() => {
-    if (Boolean(nftOpenActionEmbed)) {
-      setPublicationNftOpenActionModule(nftOpenActionEmbed);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [nftOpenActionEmbed]);
-
-  useEffect(() => {
-    if (hasSwapOpenAction || Boolean(nftOpenActionEmbed)) {
-      setHasOpenAction(true);
-    } else {
-      setHasOpenAction(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [hasSwapOpenAction, nftOpenActionEmbed]);
+  const hasSwapOpenAction =
+    openAction?.address === VerifiedOpenActionModules.Swap;
 
   if (
-    !Boolean(nftOpenActionEmbed) ||
     !urls.length ||
     attachments.length ||
     quotedPublication ||
@@ -106,7 +52,7 @@ const OpenActions: FC<OpenActionProps> = ({
   if (Boolean(nftOpenActionEmbed)) {
     return (
       <div className="relative m-5">
-        <DecentOpenAction
+        <Oembed
           nftOpenActionEmbed={nftOpenActionEmbed}
           nftOpenActionEmbedLoading={nftOpenActionEmbedLoading}
           url={urls[0]}
