+++ apps/web/src/components/Publication/OpenAction/UnknownModule/Decent/index.tsx
@@ -1,13 +1,7 @@
 import type { AnyPublication } from '@hey/lens';
 import type { OG } from '@hey/types/misc';
 
-import { HEY_API_URL, IS_MAINNET } from '@hey/data/constants';
-import { VerifiedOpenActionModules } from '@hey/data/verified-openaction-modules';
-import getFavicon from '@hey/helpers/getFavicon';
-import { isMirrorPublication } from '@hey/helpers/publicationHelpers';
-import { useQuery } from '@tanstack/react-query';
-import axios from 'axios';
-import { type FC, useEffect, useState } from 'react';
+import { type FC } from 'react';
 
 import CreatePublicationEmbed from './CreatePublicationEmbed';
 import FeedEmbed from './FeedEmbed';
@@ -23,87 +17,27 @@ export const openActionCTA = (platformName?: string): string => {
     : 'Mint';
 };
 interface DecentOpenActionProps {
+  isFullPublication?: boolean;
   mirrorPublication?: AnyPublication;
-  nftOpenActionEmbed?: boolean;
-  nftOpenActionEmbedLoading?: boolean;
-  onLoad?: (og: OG) => void;
+  og: OG;
+  openActionEmbed: boolean;
+  openActionEmbedLoading: boolean;
   publication?: AnyPublication;
-  url: string;
 }
 
 const DecentOpenAction: FC<DecentOpenActionProps> = ({
   mirrorPublication,
-  nftOpenActionEmbed,
-  nftOpenActionEmbedLoading,
-  onLoad,
-  publication,
-  url
+  og,
+  openActionEmbed,
+  openActionEmbedLoading,
+  publication
 }) => {
-  const [currentPublication, setCurrentPublication] =
-    useState<AnyPublication>();
-
-  const { data } = useQuery({
-    enabled: Boolean(url),
-    queryFn: async () => {
-      const response = await axios.get(`${HEY_API_URL}/oembed`, {
-        params: { url }
-      });
-      return response.data.oembed;
-    },
-    queryKey: ['getOembed', url],
-    refetchOnMount: false
-  });
-
-  useEffect(() => {
-    if (publication) {
-      setCurrentPublication(publication);
-    }
-  }, [publication]);
-
-  useEffect(() => {
-    if (onLoad) {
-      onLoad(data as OG);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data]);
-
-  const og: OG = {
-    description: data?.description,
-    favicon: data?.url ? getFavicon(data.url) : null,
-    html: data?.html,
-    image: data?.image,
-    nft: data?.nft,
-    site: data?.site,
-    title: data?.title,
-    url: url as string
-  };
-
-  const targetPublication =
-    currentPublication && isMirrorPublication(currentPublication)
-      ? currentPublication.mirrorOn
-      : currentPublication;
-
-  const canPerformDecentAction: boolean = Boolean(
-    targetPublication &&
-      targetPublication.openActionModules.some(
-        (module) =>
-          module.contract.address === VerifiedOpenActionModules.DecentNFT
-      )
-  );
-
-  const embedDecentOpenAction: boolean =
-    IS_MAINNET && (canPerformDecentAction || Boolean(nftOpenActionEmbed));
-
-  if (!embedDecentOpenAction) {
-    return null;
-  }
-
   if (!publication) {
     return (
       <CreatePublicationEmbed
         og={og}
-        openActionEmbed={Boolean(nftOpenActionEmbed)}
-        openActionEmbedLoading={Boolean(nftOpenActionEmbedLoading)}
+        openActionEmbed={openActionEmbed}
+        openActionEmbedLoading={openActionEmbedLoading}
       />
     );
   }
@@ -112,8 +46,8 @@ const DecentOpenAction: FC<DecentOpenActionProps> = ({
     <FeedEmbed
       mirrorPublication={mirrorPublication}
       og={og}
-      openActionEmbed={Boolean(nftOpenActionEmbed)}
-      openActionEmbedLoading={Boolean(nftOpenActionEmbedLoading)}
+      openActionEmbed={openActionEmbed}
+      openActionEmbedLoading={openActionEmbedLoading}
       publication={publication}
     />
   );
