+++ apps/web/src/components/Publication/OpenAction/UnknownModule/RentableBillboard/index.tsx
@@ -1,21 +1,9 @@
-import type {
-  MirrorablePublication,
-  UnknownOpenActionModuleSettings
-} from '@hey/lens';
 import type { FC } from 'react';
 
 import stopEventPropagation from '@hey/helpers/stopEventPropagation';
 import { Button, Card } from '@hey/ui';
 
-interface RentableBillboardOpenActionProps {
-  module: UnknownOpenActionModuleSettings;
-  publication?: MirrorablePublication;
-}
-
-const RentableBillboardOpenAction: FC<RentableBillboardOpenActionProps> = ({
-  module,
-  publication
-}) => {
+const RentableBillboardOpenAction: FC = () => {
   return (
     <Card
       className="space-y-4 p-10 text-center"
